Enc experiment (#121)
* Optimisating noise and sinusoid

* Update Noise and refactoring

* Lighters update

* Refactoring Bouncing Balls

* Refactoring Butterflys and Snake island

* Refactoring SmokeBalls

* Refactoring rings

* Update Kaleidoscope

* WRain update(test)

* update wrain(clouds)

* Optional white clouds in Wrain

* Fixing clouds and returning lost original look

* FIxing wrain control

* Shifting center in Kaleidoscope

* Fixing shifting

* Fix Builder for new PIO rules for names of targets

* Translation fixing

* Переклад, правки

* Testing encoder fixing

* Color fix for gauge (encoder)

---------

+++ include/extra_tasks.h
@@ -114,7 +114,7 @@ void GaugeMix(GAUGETYPE type = GAUGETYPE::GT_NONE) {
             for (uint8_t x = 0; x <= xCol * (xStep - 1); x += xStep) {
                 for(uint8_t y = 0; y < HEIGHT; y++){
                     EffectMath::drawPixelXY(x, y, (y < ind)? ((gauge_hue)? CHSV(gauge_hue, 255, 255) : CRGB(gauge_color)):
-                     ((ind == y && subPart)? ((gauge_hue)? CHSV(gauge_hue, 255, subPart) : CRGB(gauge_color.r*255/subPart,gauge_color.g*255/subPart,gauge_color.b*255/subPart)):
+                     ((ind == y && subPart)? ((gauge_hue)? CHSV(gauge_hue, 255, subPart) : CRGB((uint16_t)gauge_color.r * 255 / subPart, (uint16_t)gauge_color.g * 255 / subPart, (uint16_t)gauge_color.b * 255 / subPart)):
                       CRGB(0,0,0)));
                 }
             }
@@ -125,7 +125,7 @@ void GaugeMix(GAUGETYPE type = GAUGETYPE::GT_NONE) {
             for (uint8_t y = 0; y <= yCol * (yStep - 1); y += yStep) {
                 for(uint8_t x = 0; x < WIDTH; x++){
                     EffectMath::drawPixelXY((x + y) % WIDTH, y, (x < ind)? ((gauge_hue)? CHSV(gauge_hue, 255, 255) : CRGB(gauge_color)):
-                     ((ind == x && subPart)? ((gauge_hue)? CHSV(gauge_hue, 255, subPart) : CRGB(gauge_color.r*255/subPart,gauge_color.g*255/subPart,gauge_color.b*255/subPart)):
+                     ((ind == x && subPart)? ((gauge_hue)? CHSV(gauge_hue, 255, subPart) : CRGB((uint16_t)gauge_color.r * 255 / subPart, (uint16_t)gauge_color.g * 255 / subPart , (uint16_t)gauge_color.b * 255 / subPart)):
                       CHSV(0,0,0)));
                 }
             }
